Trying with two contexts, one for spaces, one for types.
This does not work nicely because we are required to speak
about points of a cover as term judgements, but these refer
to the whole type context. While for the definition of covers
we didn't want to speak about the whole type context. Once we
change the definition of covers to refer to the type context,
so does the definition of spaces. This makes the distinction
between type and space contexts vacuous.