Trying to formalize Loc/Val with a single subset annotation.
But it seems like we probably need two, because in a commtype,
if we have (0 -> 1)[Val a A] then the A has to be available at
{0,1} but the a only at {0}. And this weaker condition has to
be passed down from the communication constructor to the type
which is inside.

We are thus going to try to add a second annotation.