Trying to implement the two argument-indexed types,
but it seems like we actually do need a distinction
between [] as index and global types. Because [] as
index means that this value is implemented at no types
while a global type has to be implemented of course.

Next step: reintroduce global/local type distinction,
but with a common context.